requirements.txt
google-genai
python-dotenv
genanki